pyproject.toml
[project]
name = "bloomberg-a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Samuel Seng", email = "djsamseng@gmail.com" }
]
requires-python = ">=3.12, <3.13"
dependencies = [
    "bloomberg-mcp",
    "google-a2a",
    "langchain>=0.3.23",
    "langchain-core>=0.3.51",
    "langchain-mcp-adapters>=0.0.7",
    "langchain-ollama>=0.3.1",
    "langgraph>=0.3.29",
]

[project.scripts]
bloomberg-a2a = "bloomberg_a2a:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
name = "blpapi"
url = "https://blpapi.bloomberg.com/repository/releases/python/simple/"

[tool.uv.sources]
bloomberg-mcp = { git = "https://github.com/djsamseng/bloomberg-mcp" }
google-a2a = { git = "https://github.com/djsamseng/A2A", subdirectory = "samples/python", branch = "prefixPythonPackage" }
